api/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic>=2.6.0
python-multipart==0.0.9
requests==2.31.0
uuid>=1.30
aiofiles>=23.2.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
yfinance==0.2.36
pandas==2.2.0
numpy==1.26.3 